pyproject.toml
[tool.poetry]
name = "hoiit"
version = "0.1.0"
description = ""
authors = ["ngoctuct"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
pyTelegramBotAPI = "^4.5.1"
aiohttp = "^3.9.5"
tinydb = "^4.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
